Guys, Please dont use echo. Use print instead. Echo is meant to me a server environment call. Print is a php call.
Echo may break if the server system doesnt support it. or if your php
configuration doesnt allow the use of system calls.

Also, Devin if you want to pad string's use str_pad function. Using
a log in a for loop is very inefficient and doesnt work if the input
is not a number... OR if the number is a 0